From Yue: fix some coverity reports
git-svn-id: http://root.cern.ch/svn/root/trunk@47481 27541ba8-7e3a-0410-8455-c3a389f83636

+++ graf2d/mathtext/src/fontembed.cxx
@@ -801,7 +801,7 @@ namespace mathtext {
 			}
 #endif // LITTLE_ENDIAN
 
-			uint16_t max_glyph_name_index = 0;
+			size_t max_glyph_name_index = 0;
 			for (int i = num_glyphs - 1; i >= 0; i--) {
 				if (glyph_name_index[i] > max_glyph_name_index) {
 					max_glyph_name_index = glyph_name_index[i];
